First pass on the executing retry code
I'm planning on doing some nonsense around the SCAN lookup since the
docs for KEYS tells you not to use it.

I still need to decide if I want to use the `redis` library's iterator
support and hope we'll never have enough expired entries to overflow a
pipeline's max capacity or DIY something to work in small batches.

I'm leaning towards the DIY front purely because it makes it slightly
more robust and god knows we need robustness here